fix flickering display during high log traffic
to reduce flicker we want to avoid redraws of the terminal as much as possible.
since it seems that terminal emulators flush their display buffers immediately
when they see a screen clear command but hold on to them for further
modifications when the see a line clear command we'll just clear every line
individually and overwrite it from there. this could result in a small gap at
the end of the display if the nom table loses more rows than it receives output
lines in the same update cycle, but the same would've happened with screen clear too.