Reset console state when terminating through SIGINT/SIGTERM
Nom alter the internal state of the underlying console it's running
in. We need to reset that internal set to a suitable default before
leaving the program when killed with SIGTERM/SIGINT.

In the process of solving this issue, we introduce a custom signal
handler in charge to reset the console cursor and plug it to SIGINT
and SIGTERM. This signal handler terminates the overall program by
killing the main thread with exit code 1. The RTS runtime then kills
the other thread for us before exiting itself with code 1.

Fixes https://github.com/maralorn/nix-output-monitor/issues/114