Toggle between drvname/name when pressing n
Here's my personal use case for this:

Sometimes, I'm on my laptop and start a build before realizing I'm
about to build something massive. In that case, I'd like to copy paste
the drv path to launch the build on a beefy remote builder. The idea
is to make NOM print the drv path instead of the derivation name when
pressing the "n" key.