Freeze/help user triggered features
When the user presses h or ?, we print a short usage/help text listing
the interactive key bindings.

When the user presses f, we freeze the output text. I personally use
this to copy/paste stuff from the terminal output, that's quite
convenient. There's no visual feedback showing the output text is
paused at the moment. This might be a UX footgun ><